+++ src/impls/mod.rs
@@ -2,6 +2,7 @@ mod bevy_impls;
 mod list;
 mod number;
 mod primitives;
+mod quat;
 mod ui;
 mod vec;
 pub(crate) mod with_context;
@@ -9,6 +10,7 @@ pub(crate) mod with_context;
 pub use bevy_impls::ColorAttributes;
 pub use number::NumberAttributes;
 pub use primitives::{OptionAttributes, StringAttributes};
+pub use quat::{QuatAttributes, QuatDisplay};
 pub use vec::Vec2dAttributes;
 pub use with_context::{FilterType, TextureAttributes};